refactor: move career path creation into separate file
